requirements.txt
# Stock Query App Dependencies
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
